requirements.txt
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
tqdm>=4.65.0
jupyter>=1.0.0
neuraloperator>=0.3.0
einops>=0.7.0
tensorboard>=2.13.0
